fix: do not create data dir at each `simapp` run
